+++ src/main/java/net/causw/config/security/SecurityService.java
@@ -7,8 +7,8 @@ import net.causw.domain.aop.annotation.MeasureTime;
 import net.causw.domain.model.enums.user.RoleGroup;
 import net.causw.domain.model.enums.user.Role;
 import net.causw.domain.model.util.MessageUtil;
-import net.causw.domain.policy.RolePolicy;
-import net.causw.domain.policy.UserPolicy;
+import net.causw.domain.policy.security.RolePolicy;
+import net.causw.domain.policy.security.UserPolicy;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.authentication.AnonymousAuthenticationToken;
 import org.springframework.security.core.Authentication;
